Convenience functions for reading and writing archives from disk

If you want to simply read an archive from disk and extract some files back to disk, you can use the convenience method unzip_files or unzip_file:

using ZipStreams

# write an archive to disk
archive_name = tempname()
zipsink(archive_name) do sink
    write_file(sink, "hello.txt", "Hello, Julia!")
    write_file(sink, "subdir/goodbye.txt", "Goodbye, Julia!"; make_path=true)
    write_file(sink, "test.txt", "This is a test.")
end

# extract the contents here (".")
unzip_files(archive_name)
@assert read("hello.txt", String) == "Hello, Julia!"
@assert read("subdir/goodbye.txt", String) == "Goodbye, Julia!"
@assert read("test.txt", String) == "This is a test."

# extract files somewhere else
outdir = tempdir()
unzip_files(archive_name; output_path=outdir)
@assert read(joinpath(output_path, "hello.txt"), String) == "Hello, Julia!"
@assert read(joinpath(output_path, "subdir/goodbye.txt"), String) == "Goodbye, Julia!"
@assert read(joinpath(output_path, "test.txt"), String) == "This is a test."

# extract specific files here, making subdirectories as needed
unzip_files(archive_name, ["hello.txt", "subdir/goodbye.txt"])
@assert read(joinpath(output_path, "hello.txt"), String) == "Hello, Julia!"
@assert read(joinpath(output_path, "subdir/goodbye.txt"), String) == "Goodbye, Julia!"

# extract specific files somewhere else, making the root directory if needed
unzip_files(archive_name, "test.txt"; output_path="other/location", make_path=true)
@assert read("other/location/test.txt", String) == "This is a test."

If you want to store files from disk to a new archive, you can use the zip_files or zip_file method:

using ZipStreams

# make some fake files to compress
dir = mktempdir()
path1 = tempname(dir)
write(path1, "Hello, Julia!")
path2 = tempname(dir)
write(path2, "Goodbye, Julia!")

# Archive the files
archive_name = tempname(dir)
zip_files(archive_name, [path1, path2])

API

ZipStreams.unzip_files — Function

unzip_files(archive; output_path::AbstractString=".", make_path::Bool=false)
unzip_files(archive, files; [keyword_args])

Unzip files from archive. If files is not given, extract all files.

This method opens the archive and iterates through the archived files, writing them to disk to the directory tree rooted at output_path. If make_path is true and output_path does not exist, it will be created.

See zipsource and next_file for more information about how the files are read from the archive.

ZipStreams.zip_files — Function

zip_files(out_filename, files; [keyword_args])
zip_files(out_filename, dir; recurse_directories=false, [keyword_args])

Create an archive from files on disk.

The archive out_filename will be created using the zipsink method with the given keyword arguments. in_filename can be a single path or a vector of multiple paths on disk. The files will be written in the archive with paths matching the closest common relative path between the current directory (".") and the full path of the file, so if archive_filename is "/a/b/archive.zip" and one of files is "/a/c/file", then the file will be witten with the path "c/file".

If dir is a directory and recurse_directories is true, then all files and directories found when traversing the directory will be added to the archive. If recurse_directories is false (the default), then subdirectories of dir will not be traversed.

All files are written to the archive using the default arguments specified by open(zipsink, fn). See open(::ZipArchiveSink, ::AbstractString) for more information.

See zipsink for more information about the optional keyword arguments.